requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
backports.zoneinfo==0.2.1
bleach==6.0.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-graphql-jwt==0.3.4
djangorestframework==3.14.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
idna==3.4
Markdown==3.3.4
promise==2.3
psycopg2-binary
PyJWT==2.6.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
urllib3==1.26.14
webencodings==0.5.1
uwsgi